requirements.txt
requests
tqdm
colorama
biopython
install
psutil
py-cpuinfo